import tarball aegis-1995.03.13-02.003

+++ test/00/t0002a.sh
@@ -1,7 +1,7 @@
-#! /bin/sh
+#!/bin/sh
 #
 #	aegis - project change supervisor
-#	Copyright (C) 1991, 1992, 1993 Peter Miller.
+#	Copyright (C) 1991, 1992, 1993, 1994, 1995 Peter Miller;
 #	All rights reserved.
 #
 #	This program is free software; you can redistribute it and/or modify
@@ -23,16 +23,24 @@
 
 unset AEGIS_PROJECT
 unset AEGIS_CHANGE
+unset AEGIS_PATH
+unset AEGIS
 umask 022
 
 USER=${USER:-${LOGNAME:-`whoami`}}
 
 work=${AEGIS_TMP:-/tmp}/$$
 
+here=`pwd`
+if test $? -ne 0; then exit 1; fi
+
+if test "$1" != "" ; then bin="$here/$1/bin"; else bin="$here/bin"; fi
+
 fail()
 {
 	set +x
 	echo FAILED test of core functionality 1>&2
+	cd $here
 	find $work -type d -user $USER -exec chmod u+w {} \;
 	rm -rf $work
 	exit 1
@@ -41,6 +49,7 @@ pass()
 {
 	set +x
 	echo PASSED 1>&2
+	cd $here
 	find $work -type d -user $USER -exec chmod u+w {} \;
 	rm -rf $work
 	exit 0
@@ -52,6 +61,13 @@ trap "fail" 1 2 3 15
 #
 PAGER=cat
 export PAGER
+
+AEGIS_FLAGS="delete_file_preference = no_keep; \
+	diff_preference = automatic_merge;"
+export AEGIS_FLAGS
+AEGIS_THROTTLE=2
+export AEGIS_THROTTLE
+
 worklib=$work/lib
 workproj=$work/foo.proj
 workchan=$work/foo.chan
@@ -62,14 +78,16 @@ tmp=$work/tmp
 #
 mkdir $work
 if test $? -ne 0 ; then exit 1; fi
+cd $work
+if test $? -ne 0 ; then exit 1; fi
 
 #
 # make a new project
 #	and check files it should have made
 #
-./bin/aegis -newpro foo -dir $workproj -lib $worklib
+$bin/aegis -newpro foo -dir $workproj -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newpro -list -lib $worklib
+$bin/aegis -newpro -list -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $worklib/state ; then fail; fi
 if test ! -r $workproj/info/state ; then fail; fi
@@ -83,9 +101,9 @@ developer_may_review = true;
 developer_may_integrate = true;
 reviewer_may_integrate = true;
 end
-./bin/aegis -proatt -list --proj=foo -lib $worklib
+$bin/aegis -proatt -list --proj=foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -proatt $tmp -proj foo -lib $worklib
+$bin/aegis -proatt -f $tmp -proj foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -97,9 +115,9 @@ brief_description = "This change is used to test the aegis functionality \
 with respect to change descriptions.";
 cause = internal_bug;
 end
-./bin/aegis -newcha -list -pro foo -lib $worklib
+$bin/aegis -newcha -list -pro foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_change $tmp -project foo -lib $worklib
+$bin/aegis -new_change -f $tmp -project foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/state ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail; fi
@@ -113,45 +131,45 @@ brief_description = "This change was added to make the various listings \
 much more interesting.";
 cause = internal_bug;
 end
-./bin/aegis -new_change $tmp -project foo -lib $worklib
+$bin/aegis -new_change -f $tmp -project foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_change -list -project foo -lib $worklib
+$bin/aegis -new_change -list -project foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # add a new developer
 #
-./bin/aegis -newdev -list -pr foo -lib $worklib
+$bin/aegis -newdev -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newdev $USER -p foo -lib $worklib
+$bin/aegis -newdev $USER -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newdev -list -pr foo -lib $worklib
+$bin/aegis -newdev -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # begin development of a change
 #	check it made the files it should
 #
-./bin/aegis -devbeg -list -project foo -lib $worklib
+$bin/aegis -devbeg -list -project foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -devbeg 1 -p foo -dir $workchan -lib $worklib
+$bin/aegis -devbeg 1 -p foo -dir $workchan -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail; fi
 if test ! -r $worklib/user/$USER ; then fail; fi
-./bin/aegis -new_change -list -project foo -lib $worklib
+$bin/aegis -new_change -list -project foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # add a new files to the change
 #
-./bin/aegis -new_file -list -lib $worklib -proJ foo
+$bin/aegis -new_file -list -lib $worklib -proJ foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_file $workchan/main.c -nl -lib $worklib -Pro foo
+$bin/aegis -new_file $workchan/main.c -nl -lib $worklib -Pro foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_file $workchan/config -nl -lib $worklib -p foo
+$bin/aegis -new_file $workchan/config -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail; fi
-./bin/aegis -new_file -list -lib $worklib -p foo
+$bin/aegis -new_file -list -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 cat > $workchan/main.c << 'end'
 void
@@ -166,9 +184,9 @@ build_command = "rm -f foo; cc -o foo -D'VERSION=\"$v\"' main.c";
 history_get_command =
 	"co -u'$e' -p $h,v > $o";
 history_create_command =
-	"ci -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
+	"ci -f -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
 history_put_command =
-	"ci -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
+	"ci -f -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
 history_query_command =
 	"rlog -r $h,v | awk '/^head:/ {print $$2}'";
 
@@ -196,9 +214,9 @@ end
 #
 # create a new test
 #
-./bin/aegis -nt -l -lib $worklib -p foo
+$bin/aegis -nt -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -nt -lib $worklib -p foo
+$bin/aegis -nt -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 cat > $workchan/test/00/t0001a.sh << 'end'
 #!/bin/sh
@@ -238,42 +256,37 @@ pass
 end
 
 #
-# let the clock tick over, so the build will be happy
-#
-sleep 1
-
-#
 # build the change
 #
-./bin/aegis -build -list -lib $worklib -p foo
+$bin/aegis -build -list -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -build -nl -lib $worklib -p foo
+$bin/aegis -build -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail; fi
 
 #
 # difference the change
 #
-./bin/aegis -diff -list -lib $worklib -p foo
+$bin/aegis -diff -list -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -diff -lib $worklib -p foo
+$bin/aegis -diff -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # test the change
 #
-./bin/aegis -test -l -lib $worklib -p foo
+$bin/aegis -test -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -lib $worklib -p foo
+$bin/aegis -test -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 
 #
 # finish development of the change
 #
-./bin/aegis -dev_end -list -lib $worklib -p foo
+$bin/aegis -dev_end -list -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -dev_end -lib $worklib -p foo
+$bin/aegis -dev_end -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -282,22 +295,22 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 #
 # add a new reviewer
 #
-./bin/aegis -newrev -list -pr foo -lib $worklib
+$bin/aegis -newrev -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newrev $USER -p foo -lib $worklib
+$bin/aegis -newrev $USER -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newrev -list -pr foo -lib $worklib
+$bin/aegis -newrev -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # fail the review
 #
-./bin/aegis -review_fail -list -p foo -lib $worklib
+$bin/aegis -review_fail -list -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 cat > $tmp << 'end'
 This is a failed review comment.
 end
-./bin/aegis -review_fail $tmp -chan 1 -proj foo -lib $worklib
+$bin/aegis -review_fail -f $tmp -chan 1 -proj foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -309,14 +322,13 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 # and test
 # end finish, again
 #
-sleep 1
-./bin/aegis -build -nl -lib $worklib -p foo
+$bin/aegis -build -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -diff -nl -lib $worklib -p foo
+$bin/aegis -diff -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -nl -lib $worklib -p foo
+$bin/aegis -test -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -dev_end -lib $worklib -p foo
+$bin/aegis -dev_end -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -325,9 +337,9 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 #
 # pass the review
 #
-./bin/aegis -review_pass -list -proj foo -lib $worklib
+$bin/aegis -review_pass -list -proj foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -review_pass -chan 1 -proj foo -lib $worklib
+$bin/aegis -review_pass -chan 1 -proj foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -336,50 +348,49 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 #
 # add an integrator
 #
-./bin/aegis -newint -list -pr foo -lib $worklib
+$bin/aegis -newint -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newint $USER -p foo -lib $worklib
+$bin/aegis -newint $USER -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -newint -list -pr foo -lib $worklib
+$bin/aegis -newint -list -pr foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # start integrating
 #
-./bin/aegis -intbeg -l -p foo -lib $worklib
+$bin/aegis -intbeg -l -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -intbeg 1 -p foo -lib $worklib
+$bin/aegis -intbeg 1 -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
 if test ! -r $worklib/user/$USER ; then fail ; fi
-./bin/aegis -integrate_begin_undo -l -lib $worklib -p foo
+$bin/aegis -integrate_begin_undo -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # make sure -chdir works
 #
-./bin/aegis -cd -l -lib $worklib -p foo
+$bin/aegis -cd -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-dir=`./bin/aegis -cd fred -bl -p foo -lib $worklib`
+dir=`$bin/aegis -cd fred -bl -p foo -lib $worklib`
 if test $? -ne 0 ; then fail; fi
 if test $dir != $workproj/baseline/fred ; then fail; fi
-dir=`./bin/aegis -cd -dd -p foo -c 1 fred -lib $worklib`
+dir=`$bin/aegis -cd -dd -p foo -c 1 fred -lib $worklib`
 if test $? -ne 0 ; then fail; fi
 if test $dir != $workchan/fred ; then fail; fi
-dir=`./bin/aegis -cd -p foo -c 1 fred -lib $worklib`
+dir=`$bin/aegis -cd -p foo -c 1 fred -lib $worklib`
 if test $? -ne 0 ; then fail; fi
 if test $dir != $workproj/delta.001/fred ; then fail; fi
 
 #
 # integrate build
 #
-./bin/aegis -build -l -lib $worklib -p foo
+$bin/aegis -build -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-sleep 1
-./bin/aegis -build -nl -lib $worklib -p foo
+$bin/aegis -build -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -nl -lib $worklib -p foo
+$bin/aegis -test -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test -f $workproj/delta.001/aegis.log ; then fail; fi
@@ -387,12 +398,12 @@ if test -f $workproj/delta.001/aegis.log ; then fail; fi
 #
 # fail the integration
 #
-./bin/aegis -intfail -l -lib $worklib -p foo
+$bin/aegis -intfail -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 cat > $tmp << 'end'
 This is a failed integration comment.
 end
-./bin/aegis -intfail $tmp -lib $worklib -p foo
+$bin/aegis -intfail -f $tmp -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -407,18 +418,17 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 # and review pass again
 # and start integrating again
 #
-sleep 1
-./bin/aegis -build -nl -lib $worklib -p foo
+$bin/aegis -build -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -diff -nl -lib $worklib -p foo
+$bin/aegis -diff -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -nl -lib $worklib -p foo
+$bin/aegis -test -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -dev_end -lib $worklib -p foo
+$bin/aegis -dev_end -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -review_pass -chan 1 -proj foo -lib $worklib
+$bin/aegis -review_pass -chan 1 -proj foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -intbeg 1 -p foo -lib $worklib
+$bin/aegis -intbeg 1 -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -427,19 +437,18 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 #
 # build and test the integration
 #
-sleep 1
-./bin/aegis -build -nl -lib $worklib -p foo
+$bin/aegis -build -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -nl -lib $worklib -p foo
+$bin/aegis -test -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 
 #
 # pass the integration
 #
-./bin/aegis -intpass -l -lib $worklib -p foo
+$bin/aegis -intpass -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -intpass -nl -lib $worklib -p foo
+$bin/aegis -intpass -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/001 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -448,23 +457,23 @@ if test ! -r $worklib/user/$USER ; then fail ; fi
 #
 # check out the listings
 #
-./bin/aegis -list -list -lib $worklib
+$bin/aegis -list -list -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -list chahist -c 1 -p foo -lib $worklib
+$bin/aegis -list chahist -c 1 -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # start work on change 2
 #
-./bin/aegis -devbeg 2 -p foo -dir $workchan -lib $worklib
+$bin/aegis -devbeg 2 -p foo -dir $workchan -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -cp -l -lib $worklib -p foo
+$bin/aegis -cp -l -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # copy a file into the change
 #
-./bin/aegis -cp $workchan/main.c -nl -lib $worklib -p foo
+$bin/aegis -cp $workchan/main.c -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/002 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
@@ -495,7 +504,7 @@ end
 #
 # need another test
 #
-./bin/aegis -nt -lib $worklib -p foo
+$bin/aegis -nt -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 cat > $workchan/test/00/t0002a.sh << 'end'
 #!/bin/sh
@@ -535,22 +544,17 @@ pass
 end
 
 #
-# tick over clock to keep build happy
-#
-sleep 1
-
-#
 # build the change
 # diff the change
 # test the change
 #
-./bin/aegis -b -nl -lib $worklib -p foo
+$bin/aegis -b -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -diff -nl -lib $worklib -p foo
+$bin/aegis -diff -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -nl -lib $worklib -p foo
+$bin/aegis -test -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -test -bl -nl -lib $worklib -p foo
+$bin/aegis -test -bl -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workchan/main.c,D ; then fail ; fi
 
@@ -560,11 +564,11 @@ if test ! -r $workchan/main.c,D ; then fail ; fi
 # review pass
 # start integrating
 #
-./bin/aegis -devend -lib $worklib -p foo
+$bin/aegis -devend -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -revpass -c 2 -p foo -lib $worklib
+$bin/aegis -revpass -c 2 -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -intbeg -c 2 -p foo -lib $worklib
+$bin/aegis -intbeg -c 2 -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -572,24 +576,23 @@ if test $? -ne 0 ; then fail; fi
 # test the integration
 # test the integration against the baseline
 #
-sleep 1
-./bin/aegis -b -nl -lib $worklib -p foo
+$bin/aegis -b -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -t -nl -lib $worklib -p foo
+$bin/aegis -t -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -t -bl -nl -lib $worklib -p foo
+$bin/aegis -t -bl -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # pass the integration
 #	make sure it create the files, etc
 #
-./bin/aegis -intpass -nl -lib $worklib -p foo
+$bin/aegis -intpass -nl -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 if test ! -r $workproj/info/change/0/002 ; then fail ; fi
 if test ! -r $workproj/info/state ; then fail ; fi
 if test ! -r $worklib/user/$USER ; then fail ; fi
-./bin/aegis -l projhist -p foo -lib $worklib
+$bin/aegis -l projhist -p foo -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 # should be no automatic logging
