+++ test/00/t0004a.sh
@@ -1,7 +1,7 @@
-#! /bin/sh
+#!/bin/sh
 #
 #	aegis - project change supervisor
-#	Copyright (C) 1991, 1992, 1993 Peter Miller.
+#	Copyright (C) 1991, 1992, 1993, 1994, 1995 Peter Miller;
 #	All rights reserved.
 #
 #	This program is free software; you can redistribute it and/or modify
@@ -28,16 +28,24 @@
 
 unset AEGIS_PROJECT
 unset AEGIS_CHANGE
+unset AEGIS_PATH
+unset AEGIS
 umask 022
 
 USER=${USER:-${LOGNAME:-`whoami`}}
 
 work=${AEGIS_TMP:-/tmp}/$$
 
+here=`pwd`
+if test $? -ne 0; then exit 1; fi
+
+if test "$1" != "" ; then bin="$here/$1/bin"; else bin="$here/bin"; fi
+
 fail()
 {
 	set +x
 	echo "FAILED test of 'aegis -DIFFerence' command" 1>&2
+	cd $here
 	find $work -type d -user $USER -exec chmod u+w {} \;
 	rm -rf $work
 	exit 1
@@ -46,6 +54,7 @@ pass()
 {
 	set +x
 	echo PASSED 1>&2
+	cd $here
 	find $work -type d -user $USER -exec chmod u+w {} \;
 	rm -rf $work
 	exit 0
@@ -57,6 +66,13 @@ trap "fail" 1 2 3 15
 #
 PAGER=cat
 export PAGER
+
+AEGIS_FLAGS="delete_file_preference = no_keep; \
+	diff_preference = automatic_merge;"
+export AEGIS_FLAGS
+AEGIS_THROTTLE=2
+export AEGIS_THROTTLE
+
 worklib=$work/lib
 workproj=$work/foo.proj
 workchan=$work/foo.chan
@@ -71,11 +87,14 @@ set -x
 # make the directories
 #
 mkdir $work
+if test $? -ne 0 ; then fail; fi
+cd $work
+if test $? -ne 0 ; then fail; fi
 
 #
 # make a new project
 #
-./bin/aegis -newpro foo -dir $workproj -v -lib $worklib
+$bin/aegis -newpro foo -dir $workproj -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -87,7 +106,7 @@ developer_may_review = true;
 developer_may_integrate = true;
 reviewer_may_integrate = true;
 end
-./bin/aegis -proatt $tmp -proj foo -v -lib $worklib
+$bin/aegis -proatt -f $tmp -proj foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -99,7 +118,7 @@ brief_description = "This change is used to test the aegis functionality \
 with respect to change descriptions.";
 cause = internal_bug;
 end
-./bin/aegis -new_change $tmp -project foo -v -lib $worklib
+$bin/aegis -new_change -f $tmp -project foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -111,29 +130,29 @@ brief_description = "This change was added to make the various listings \
 much more interesting.";
 cause = internal_bug;
 end
-./bin/aegis -new_change $tmp -project foo -v -lib $worklib
+$bin/aegis -new_change -f $tmp -project foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # add a new developer
 #
-./bin/aegis -newdev $USER -p foo -v -lib $worklib
+$bin/aegis -newdev $USER -p foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # begin development of a change
 #
-./bin/aegis -devbeg 1 -p foo -dir $workchan -v -lib $worklib
+$bin/aegis -devbeg 1 -p foo -dir $workchan -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # add a new files to the change
 #
-./bin/aegis -new_file $workchan/main.c -nl -v -lib $worklib -p foo
+$bin/aegis -new_file $workchan/main.c -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_file $workchan/fubar -nl -v -lib $worklib -p foo
+$bin/aegis -new_file $workchan/fubar -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
-./bin/aegis -new_file $workchan/config -nl -v -lib $worklib -p foo
+$bin/aegis -new_file $workchan/config -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -157,9 +176,9 @@ link_integration_directory = true;
 history_get_command =
 	"co -u'$e' -p $h,v > $o";
 history_create_command =
-	"ci -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
+	"ci -f -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
 history_put_command =
-	"ci -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
+	"ci -f -u -m/dev/null -t/dev/null $i $h,v; rcs -U $h,v";
 history_query_command =
 	"rlog -r $h,v | awk '/^head:/ {print $$2}'";
 
@@ -172,7 +191,7 @@ end
 #
 # create a new test
 #
-./bin/aegis -nt -v -lib $worklib -p foo
+$bin/aegis -nt -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -216,93 +235,87 @@ pass
 end
 
 #
-# let the clock tick over, so the build will be happy
-#
-sleep 1
-
-#
 # build the change
 #
-./bin/aegis -build -nl -v -lib $worklib -p foo
+$bin/aegis -build -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # difference the change
 #
-./bin/aegis -diff -nl -v -lib $worklib -p foo
+$bin/aegis -diff -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # test the change
 #
-./bin/aegis -test -nl -v -lib $worklib -p foo
+$bin/aegis -test -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # finish development of the change
 #
-./bin/aegis -dev_end -v -lib $worklib -p foo
+$bin/aegis -dev_end -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # add a new reviewer
 #
-./bin/aegis -newrev $USER -p foo -v -lib $worklib
+$bin/aegis -newrev $USER -p foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # pass the review
 #
-./bin/aegis -review_pass -chan 1 -proj foo -v -lib $worklib
+$bin/aegis -review_pass -chan 1 -proj foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # add an integrator
 #
-./bin/aegis -newint $USER -p foo -v -lib $worklib
+$bin/aegis -newint $USER -p foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # start integrating
 #
-./bin/aegis -intbeg 1 -p foo -v -lib $worklib
+$bin/aegis -intbeg 1 -p foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # integrate build
 #
-sleep 1
-./bin/aegis -build -nl -v -lib $worklib -p foo
+$bin/aegis -build -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # integrate test
 #
-./bin/aegis -test -nl -v -lib $worklib -p foo
+$bin/aegis -test -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # integration pass
 #
-./bin/aegis -intpass -nl -v -lib $worklib -p foo
+$bin/aegis -intpass -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # start work on change 2
 #
-./bin/aegis -devbeg 2 -p foo -v -dir $workchan -lib $worklib
+$bin/aegis -devbeg 2 -p foo -v -dir $workchan -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
 # copy a file into the change
 #
-./bin/aegis -cp $workchan/main.c -nl -v -lib $worklib -p foo
+$bin/aegis -cp $workchan/main.c -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 #
 # remove a file with the change
 #
-./bin/aegis -rm $workchan/fubar -nl -p foo -v -lib $worklib
+$bin/aegis -rm $workchan/fubar -nl -p foo -v -lib $worklib
 if test $? -ne 0 ; then fail; fi
 
 #
@@ -330,7 +343,7 @@ end
 #
 # need another test
 #
-./bin/aegis -nt -v -lib $worklib -p foo
+$bin/aegis -nt -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 cat > $workchan/test/00/t0002a.sh << 'end'
 #!/bin/sh
@@ -370,15 +383,10 @@ pass
 end
 
 #
-# tick over clock to keep build happy
-#
-sleep 1
-
-#
 # diff the change
 #	This is what all the set-up was about!
 #
-./bin/aegis -diff -nl -v -lib $worklib -p foo
+$bin/aegis -diff -nl -v -lib $worklib -p foo
 if test $? -ne 0 ; then fail; fi
 
 # should be no automatic logging
